Rename provider from 'nextgen' to 'native'
