+++ localstack-core/localstack/services/kinesis/native/models.py
